requirements.txt
absl-py==1.0.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
cycler==0.11.0
fonttools==4.32.0
GDAL @ file:///D:/wu/GDAL-3.4.2-cp37-cp37m-win_amd64.whl
google-auth==2.6.5
google-auth-oauthlib==0.4.6
grpcio==1.44.0
idna==3.3
importlib-metadata==4.11.3
kiwisolver==1.4.2
Markdown==3.3.6
matplotlib==3.5.1
numpy==1.21.6
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
python-dateutil==2.8.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.7.3
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch @ file:///C:/Users/Administrator/Downloads/torch-1.8.0-cp37-cp37m-win_amd64.whl
torchaudio @ file:///C:/Users/Administrator/Downloads/torchaudio-0.8.0-cp37-none-win_amd64.whl
torchvision @ file:///C:/Users/Administrator/Downloads/torchvision-0.9.0-cp37-cp37m-win_amd64.whl
tqdm==4.64.0
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
wincertstore==0.2
zipp==3.8.0
